Given an AutoValue property like `ImmutableList<String> strings()`, allow the AutoValue builder to define both `setStrings(ImmutableList<String>)` and `stringsBuilder()`. At runtime, if `stringsBuilder` is called then it is an error to call `setStrings` afterwards. Previously, it was a compile-time error just to define both methods.
In addition to the semantic change, there is a performance optimization. Previously if you made a builder starting from an existing @AutoValue instance, and constructed a second instance, the second instance's ImmutableList<String> would be a copy of the original even if it was never changed. Now, if you don't call stringsBuilder() then the second instance's ImmutableList<String> will be the same object as the first instance's.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=107052717